More work on export dialog.

+++ src/main/scala/moe/lymia/princess/renderer/manager.scala
@@ -36,7 +36,7 @@ case class SVGData(private val builder: SVGBuilder, private val definition: SVGD
   val size = builder.settings.size
 
   private def norm(d: Double) = math.max(1, math.round(d)).toInt
-  def bestSizeForDPI(dpi: Int) =
+  def bestSizeForDPI(dpi: Double) =
     (norm(size.width  / size.unit.unPerInch * dpi),
      norm(size.height / size.unit.unPerInch * dpi))
